DocIQ/requirements.txt
# DocIQ — Requirements
# Install: pip install -r requirements.txt

# ── Core Framework ──
streamlit>=1.35.0

# ── LLM Providers ──
anthropic>=0.28.0
openai>=1.30.0          # optional OpenAI backend

# ── Embeddings ──
sentence-transformers>=3.0.0
torch>=2.0.0             # CPU build is fine; GPU optional

# ── Vector Database ──
faiss-cpu>=1.8.0         # use faiss-gpu for GPU acceleration

# ── Document Parsing ──
pdfplumber>=0.11.0       # primary PDF parser
PyPDF2>=3.0.0            # PDF fallback
python-docx>=1.1.0       # DOCX parsing

# ── Token Counting ──
tiktoken>=0.7.0          # OpenAI tokenizer (cl100k_base)

# ── Evaluation ──
rouge-score>=0.1.2       # ROUGE-1/2/L metrics

# ── Utilities ──
numpy>=1.26.0
requests>=2.31.0         # Ollama HTTP calls
python-dotenv>=1.0.0     # .env support
